@@ -1,11 +1,11 @@
 {
-  "name": "@codemod/core",
+  "name": "@codemod-esm/core",
+  "type": "module",
   "version": "2.2.0",
   "description": "Runs babel plugins for codemods, i.e. by preserving formatting using Recast.",
-  "repository": "https://github.com/codemod-js/codemod",
-  "license": "Apache-2.0",
   "author": "Brian Donovan",
-  "type": "module",
+  "license": "Apache-2.0",
+  "repository": "https://github.com/codemod-js/codemod",
   "exports": {
     "bun": "./src/index.ts",
     "default": "./dist/index.js"
@@ -21,25 +21,20 @@
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "prepublishOnly": "pnpm clean && pnpm build",
-    "test": "is-ci test:coverage test:watch",
-    "test:coverage": "bun test --coverage",
-    "test:watch": "bun test --watch"
+    "test": "bun test"
   },
   "dependencies": {
-    "@babel/core": "esm",
-    "@babel/generator": "esm",
-    "@codemod/parser": "workspace:^",
-    "is-ci-cli": "^2.2.0",
-    "recast": "^0.19.0",
-    "resolve": "^1.22.1"
+    "@babel/core": "^7.26.0",
+    "@babel/generator": "^7.26.3",
+    "@codemod-esm/parser": "workspace:^",
+    "recast": "^0.23.9"
   },
   "devDependencies": {
-    "@babel/types": "esm",
-    "@types/babel__core": "^7.1.16",
-    "@types/jest": "^29",
-    "@types/node": "^20",
-    "@types/resolve": "^1.14.0",
-    "typescript": "^5.3"
+    "@babel/types": "^7.26.3",
+    "@types/babel__core": "^7.20.5",
+    "@types/bun": "^1.1.14",
+    "@types/node": "^20.17.9",
+    "typescript": "^5.7.2"
   },
   "publishConfig": {
     "access": "public"
