@@ -4,11 +4,11 @@ export class Matcher<T> {
   }
 
   matchValue(
-    /* eslint-disable @typescript-eslint/no-unused-vars */
     value: unknown,
     keys: ReadonlyArray<PropertyKey>,
-    /* eslint-enable @typescript-eslint/no-unused-vars */
+
   ): value is T {
+    void keys
     throw new Error(`${this.constructor.name}#matchValue is not implemented`)
   }
 }
