@@ -1,6 +1,6 @@
 import type { NodePath } from '@babel/traverse'
-import * as t from '@babel/types'
-import * as m from '../matchers'
+import type * as t from '@babel/types'
+import type * as m from '../matchers'
 
 export type CapturedNodePaths<C> = {
   [K in keyof C]: C[K] extends t.Node ? NodePath<C[K]> : C[K]
@@ -14,7 +14,7 @@ export type CapturedMatchers<C> = { [K in keyof C]: m.CapturedMatcher<C[K]> }
  *
  * @example
  *
- * import * as m from '@codemod/matchers';
+ * import * as m from '@codemod-esm/matchers';
  * import { PluginObj } from '@babel/core';
  *
  * const negativeOneArgument = m.capture(m.numericLiteral(-1));
@@ -56,7 +56,7 @@ export function matchPath<Node extends t.Node, C extends m.CaptureBase>(
   callback: (paths: CapturedNodePaths<C>) => void,
 ): void {
   const toMatch = Array.isArray(value)
-    ? value.map((element) => element.node)
+    ? value.map(element => element.node)
     : value.node
   if (matcher.match(toMatch)) {
     const capturedPaths = {} as CapturedNodePaths<C>
@@ -86,17 +86,18 @@ function extractCapturedPath<C extends m.CaptureBase>(
   for (const [i, key] of keys.entries()) {
     if (typeof key === 'string') {
       if (Array.isArray(capturedPath)) {
-        throw new Error(
+        throw new TypeError(
           `failed to get '${keys.join('.')}'; at '${keys
             .slice(0, i + 1)
             .join('.')}' expected a NodePath but got an array`,
         )
       }
 
-      capturedPath = capturedPath.get(key as string)
-    } else if (typeof key === 'number') {
+      capturedPath = capturedPath.get(key)
+    }
+    else if (typeof key === 'number') {
       if (!Array.isArray(capturedPath)) {
-        throw new Error(
+        throw new TypeError(
           `failed to get '${keys.join('.')}'; at '${keys
             .slice(0, i + 1)
             .join('.')}' expected an array but got a NodePath`,
@@ -104,8 +105,9 @@ function extractCapturedPath<C extends m.CaptureBase>(
       }
 
       capturedPath = capturedPath[key]
-    } else {
-      throw new Error(
+    }
+    else {
+      throw new TypeError(
         `failed to get '${keys.join('.')}'; key '${String(
           key,
         )}' is neither a string nor a number, not ${typeof key}`,
@@ -114,10 +116,13 @@ function extractCapturedPath<C extends m.CaptureBase>(
   }
 
   if (!Array.isArray(capturedPath) && typeof capturedPath.node !== 'object') {
+    // eslint-disable-next-line ts/no-unsafe-return
     return capturedPath.node as C[keyof C] extends t.Node
       ? NodePath<C[keyof C]>
       : C[keyof C]
-  } else {
+  }
+  else {
+    // eslint-disable-next-line ts/no-unsafe-return
     return capturedPath as C[keyof C] extends t.Node
       ? NodePath<C[keyof C]>
       : C[keyof C]
