@@ -1,5 +1,5 @@
-import { promises as fs } from 'fs'
-import { rebuild, MATCHERS_FILE_PATH } from './utils/rebuild'
+import { promises as fs } from 'node:fs'
+import { MATCHERS_FILE_PATH, rebuild } from './utils/rebuild'
 
 export async function main(): Promise<number> {
   await fs.writeFile(MATCHERS_FILE_PATH, await rebuild(), 'utf8')
