@@ -1,9 +1,9 @@
 import * as Babel from '@babel/core'
 import * as t from '@babel/types'
 
-export * from './NodeTypes'
 export * from './builders'
 export * from './js'
 export * from './nodesEquivalent'
+export * from './NodeTypes'
 
 export { Babel, t, t as types }
